rename FileIndexer to JournalConceptIndexer, improve its indexing.
