+++ dspace/config/dspace.cfg
@@ -2605,7 +2605,7 @@ solr.log.dataonemn.server = ${default.solr.log.dataonemn.server}
 
 # LOCAL AUTHORITY CONTROLLED FIELD: solr is populated by the class LocalIndexer that refers to every single field reported below.
 # URL to reach SOLR
-solr.authority.indexes=prism.publicationName
+#solr.authority.indexes=prism.publicationName
 
 
 choices.plugin.prism.publicationName =  SolrAuthorityType
