+++ dspace/modules/api/src/main/resources/spring/spring-dspace-core-services.xml
@@ -20,10 +20,9 @@
 
     <bean class="org.dspace.identifier.DOIIdentifierProvider" scope="singleton"/>
 
-    <bean id="com.atmire.authority.indexer.FileIndexer" class="org.datadryad.authority.FileIndexer"/>
-
-
     <bean id="com.atmire.authority.indexer.LocalIndexer" class="org.datadryad.authority.LocalIndexer"/>
+
+    <bean id="com.atmire.authority.indexer.JournalConceptIndexer" class="org.datadryad.authority.JournalConceptIndexer"/>
 
 
 </beans>
